Add deeplink recipe for single module A recipe that shows how to parse a requested deeplink into a backStack key. It demonstrates a few crucial steps:
STEP 1.Parse supported deeplinks (URLs that can be deeplinked into) into a readily readable format (see [DeepLinkCandidate])
STEP 2. Parse the requested deeplink into a readily readable, format (see [DeepLinkTarget])
 **note** the parsed requested deeplink and parsed supported deeplinks should be cohesive with each other to facilitate comparison and finding a match
STEP 3. Compare the requested deeplink target with supported deeplinks in order to find a match (see [DeepLinkMatchResult]). The match result's format should enable conversion from result to backstack key, regardless of what the conversion method may be.
STEP 4. Associate the match results with the correct backstack key

This recipes provides an example for each of the above steps by way of kotlinx.serialization.

Note that this recipe is designed to focus on parsing an intent into a key, and therefore the following deeplink considerations are not included in this scope:
 - Create synthetic backStack
 - Multi-modular setup
 - DI
 - Managing TaskStack
 - Up button ves Back Button