Implement Advanced Deeplink Recipe
The recipe spans across two modules in order to simulate a real-world scenario where `App A` deeplinks into `App B`.

The goal of this recipe is to show how to:

1. Create a synthetic backStack

A synthetic backStack is created when the deeplinked app is started as the root Activity of a new Task. This backStack is required
in order to support proper Back & Up buttons.

2. Support Back / Up buttons

Back button brings user back to the prevoius screen, while the Up button brings user to the current screen's hierarhical parent. This requires
synthetic backStack and Task stack management. See README for more info.